+++ lib/redux_future.dart
@@ -3,20 +3,24 @@ library redux_future;
 import 'dart:async';
 import 'package:redux/redux.dart';
 
-/// If you'd like to dispatch the result of a [Future] to your Redux [Store],
-/// without needing access to the [Store.dispatch] method directly in a Widget,
-/// you can attach the [futureMiddleware] to your [Store] and dispatch a
-/// simple [Future], or [FutureAction] that contains your [Future] inside.
+/// A Redux [Store] Middleware for handling Dart Futures as actions, with
+/// support for optimistic payloads.
 ///
-/// The [Future] or [FutureAction] will be intercepted by the middleware. If the
-/// future completes successfully, a [FutureFulfilledAction] will be dispatched
-/// with the result of the future. If the future fails, a [FutureRejectedAction]
-/// will be dispatched containing the error that was returned.
+/// The `futureMiddleware` can be attached to the Redux `Store` upon
+/// construction.
+///
+/// Once attached, you can `store.dispatch` a `Future` or `FutureAction`, and
+/// the `futureMiddleware` will intercept it.
+///
+///   * If the `Future` / `FutureAction` completes successfully, a
+///   `FutureFulfilledAction` will be dispatched with the result of the future.
+///   * If the `Future` / `FutureAction` fails, a `FutureRejectedAction` will be
+///   dispatched containing the error that was returned.
 ///
 /// ### Examples
 ///
-///     // First, create a reducer that knows how to handle the FutureActions:
-///     // `FutureFulfilledAction` and `FutureRejectedAction`.
+///     // First, create a reducer that knows how to handle the
+///     // `FutureFulfilledAction` and `FutureRejectedAction` actions.
 ///     String exampleReducer(String state, action) {
 ///       if (action is String) {
 ///         return action;
@@ -58,7 +62,7 @@ void futureMiddleware<State>(Store<State> store, action, NextDispatcher next) {
       next(action.initialAction);
     }
 
-    _dispatchResults(store, action.future);
+    _dispatchResults(store, action.future).then(action.completer.complete);
   } else if (action is Future) {
     _dispatchResults(store, action);
   } else {
@@ -67,20 +71,40 @@ void futureMiddleware<State>(Store<State> store, action, NextDispatcher next) {
 }
 
 // Dispatches the result of a future to the Store.
-void _dispatchResults<State>(Store<State> store, Future<dynamic> future) {
-  future
-      .then((result) => store.dispatch(new FutureFulfilledAction(result)))
-      .catchError((error) => store.dispatch(new FutureRejectedAction(error)));
+Future<FutureResultAction> _dispatchResults<State>(
+  Store<State> store,
+  Future<FutureResultAction> future,
+) {
+  return future.then((result) {
+    final fulfilledAction = new FutureFulfilledAction(result);
+    store.dispatch(fulfilledAction);
+    return fulfilledAction;
+  }).catchError((error) {
+    final errorAction = new FutureRejectedAction(error);
+    store.dispatch(errorAction);
+    return errorAction;
+  });
 }
 
-/// If you'd like to dispatch the result of a [Future] to your Redux [Store],
-/// without needing access to the [Store.dispatch] method directly, you can
-/// attach the [futureMiddleware] to your [Store] and dispatch a `FutureAction`.
+/// An Action that contains a `Future`, supports loading actions / "optimistic
+/// updates" by firing an [initialAction] before the Future completes, and can
+/// be awaited for the final FulFilled or Rejected [result] from the Middleware.
 ///
-/// The FutureAction will be intercepted by the [futureMiddleware]. If the
-/// future completes successfully, a [FutureFulfilledAction] will be dispatched
-/// with the result of the future. If the future fails, a [FutureRejectedAction]
-/// will be dispatched containing the error that was returned.
+///   * If the `Future` contained within the `FutureAction` completes
+///   successfully, a `FutureFulfilledAction` will be dispatched with the result
+///   of the future.
+///   * If the `Future` contained within the `FutureAction` fails, a
+///   `FutureRejectedAction` will be dispatched containing the error that was
+///   returned.
+///
+/// In some cases, you may want to fire an action before the Future completes.
+/// For example, you could fire a `SearchResultsFetching` action so your State
+/// can set a `fetching` value to true, and your UI can display a loading screen
+/// in response. In this case, you can use the [initialAction] parameter.
+///
+/// In addition, if you need to wait until the Fulfilled or Rejected action has
+/// been dispatched to the Store, you can wait for the [result] `Future` to
+/// complete.
 ///
 /// ### Examples
 ///
@@ -102,7 +126,19 @@ class FutureAction<T> {
 
   /// Optional: If an initialAction is provided, it will be dispatched
   /// immediately, before any actions from the [future] are returned.
+  ///
+  /// Useful for Loading Screens and "optimistic updates"
   final dynamic initialAction;
+
+  /// Internal use only. To know when the Future has completed and the
+  /// middleware has dispatched the Success or Error to the Store, use the
+  /// [result] getter.
+  final completer = new Completer<FutureResultAction>();
+
+  /// Returns a [FutureFulfilledAction] or [FutureRejectedAction] action from
+  /// the [futureMiddleware] after the middleware has dispatched the action
+  /// to the [Store].
+  Future<FutureResultAction> get result => completer.future;
 
   FutureAction(
     this.future, {
@@ -122,13 +158,16 @@ class FutureAction<T> {
 
   @override
   String toString() {
-    return 'AsyncAction{initialAction: $initialAction, future: $future}';
+    return 'FutureAction{future: $future, initialAction: $initialAction}';
   }
 }
 
+/// A base class for the Dispatched Actions.
+abstract class FutureResultAction {}
+
 /// This action will be dispatched if the [Future] provided to a [FutureAction]
 /// completes successfully.
-class FutureFulfilledAction<T> {
+class FutureFulfilledAction<T> implements FutureResultAction {
   final T result;
 
   FutureFulfilledAction(this.result);
@@ -151,7 +190,7 @@ class FutureFulfilledAction<T> {
 
 /// This action will be dispatched if the [Future] provided to a [FutureAction]
 /// finishes with an error.
-class FutureRejectedAction<E> {
+class FutureRejectedAction<E> implements FutureResultAction {
   final E error;
 
   FutureRejectedAction(this.error);